I2C communication between BL600 (Master) and main MCU is working. BLE link state: BL600 -> main MCU Cumulative Wheel Revolutions and Last Wheel Event Time (CSC Profile): main MCU -> BL600

+++ Sources/powermgr.c
@@ -53,7 +53,7 @@
 volatile bool sleep_wakeup = FALSE;
 volatile bool standby = FALSE;
 volatile bool low_energy = FALSE;
-volatile int waitTimeout = 0;
+volatile uint8_t waitTimeout = 0;
 
 // Local Variables
 // ***************
